Viper2 modifications :   * Inserting event in their right place (in scheduler.addEvent())   * Wait for next event instead of consuming all cpu (in scheduler.star())   * Adding some devices (LCD, Robot, CAN (with CAN Network), BP, DAC (in an oscilloscope))   * Modifying can driver to be able to add several CAN devices in one Ecu   * Adding can in libraries   * Inserting pygame in Viper2 (TODO : let the choice to the user to enable pygame module) and widgets to let the user communicate with Trampoline applications

+++ viper2/ipc/com.h
@@ -11,7 +11,7 @@
 
 #include "util/fifo.h"
 
-typedef int64_t reg_t; /* a register do 4 octets */
+typedef int32_t reg_t; /* a register do 4 octets */
 typedef int64_t dev_id_t;
 
 /**
